refactor: move away from monorepo approach
