+++ build-theme.js
@@ -1,7 +1,7 @@
 const postcss = require('postcss');
 const presetEnv = require('postcss-preset-env');
 const path = require('path');
-const { colors, troikaAliases, textColors } = require('./packages/tailwind-config/colors');
+const { colors, troikaAliases, textColors } = require('./src/tailwind/colors');
 
 function hexToRgb(hex) {
     var result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
