--- lerna.json
@@ -1,5 +1,0 @@
-{
-  "packages": ["packages/*"],
-  "version": "independent",
-  "useWorkspaces": true
-}
