--- packages/component-classes/.gitignore
@@ -1,1 +1,0 @@
-index.d.ts
