--- packages/component-classes/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@fabric-ds/component-classes",
-  "version": "0.0.38",
-  "type": "module",
-  "main": "index.js",
-  "types": "./index.d.ts",
-  "exports": "./index.js",
-  "author": "Dave Honneffer <pearofducks@gmail.com>",
-  "license": "MIT",
-  "scripts": {
-    "preversion": "tsc"
-  },
-  "files": [
-    "index.js",
-    "index.d.ts"
-  ],
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public"
-  },
-  "gitHead": "7b7391f45aa2c2256f0ca89ff2d5133ec34bda9c",
-  "devDependencies": {
-    "@semantic-release/changelog": "^6.0.1",
-    "@semantic-release/git": "^10.0.1",
-    "semantic-release": "^19.0.2",
-    "typescript": "^4.5.5"
-  }
-}
