--- packages/component-classes/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "allowJs": true,
-    "moduleResolution": "node",
-    "strictNullChecks": false,
-    "emitDeclarationOnly": true,
-    "declaration": true,
-    "outDir": ".",
-    "removeComments": false
-  },
-  "files": [
-    "./index.js"
-  ],
-  "exclude": [
-    "node_modules",
-  ]
-}
