--- packages/css/README.md
@@ -1,20 +1,0 @@
-# Fabric CSS
-
-This is the core package for Fabric CSS, FINNs design system.
-
-The package includes resets, core components and [Tailwind](https://tailwindcss.com/) utility classes tuned
-for FINN.
-
-## Usage (preferred)
-
-The CSS should be used directly from Eik. This is because we can alias releases by major versions.
-
-```html
-<link href="https://assets.finn.no/pkg/@fabric-ds/css/v0/fabric.min.css" rel="stylesheet" />
-```
-
-## Usage (npm)
-
-```sh
-npm install @fabric-ds/css
-```
