--- packages/css/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "@fabric-ds/css",
-  "version": "1.1.1",
-  "description": "Fabric CSS core components and utilities",
-  "files": [
-    "dist"
-  ],
-  "style": "dist/fabric.min.css",
-  "scripts": {
-    "eik:login": "npx @eik/cli login",
-    "eik:publish": "npx @eik/cli publish",
-    "eik:alias": "eik package-alias @fabric-ds/css",
-    "build": "NODE_ENV=production postcss src/fabric.css --o dist/fabric.min.css",
-    "postbuild": "./scripts/replaceTailwindVariables.mjs"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "@fabric-ds/icons": "^0.3.8",
-    "@fabric-ds/tailwind-config": "^0.5.32",
-    "@semantic-release/changelog": "^6.0.1",
-    "@semantic-release/git": "^10.0.1",
-    "semantic-release": "^19.0.2"
-  },
-  "eik": {
-    "server": "https://assets.finn.no",
-    "files": "dist"
-  },
-  "gitHead": "7b7391f45aa2c2256f0ca89ff2d5133ec34bda9c"
-}
