--- packages/tailwind-config/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@fabric-ds/tailwind-config",
-  "version": "0.5.32",
-  "description": "Fabric's Tailwind configuration",
-  "main": "tailwind.config.js",
-  "dependencies": {
-    "@tailwindcss/aspect-ratio": "^0.4.0",
-    "tailwindcss": "^2.2.19"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:fabric-ds/css.git"
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public"
-  },
-  "license": "MIT",
-  "gitHead": "7b7391f45aa2c2256f0ca89ff2d5133ec34bda9c",
-  "devDependencies": {
-    "@semantic-release/changelog": "^6.0.1",
-    "@semantic-release/git": "^10.0.1",
-    "semantic-release": "^19.0.2"
-  }
-}
