+++ pages/grid.html
@@ -27,45 +27,45 @@
           <li>
             <a href="https://css-tricks.com/snippets/css/complete-guide-grid/">A Complete Guide to Grid</a>
           </li>
-          <li><a href="https://tailwindcss.com/docs/grid-auto-flow">Grid in Tailwind</a></li>
+          <li><a href="https://tailwindcss.com/pages/grid-auto-flow">Grid in Tailwind</a></li>
         </ul>
 
         <h2>Common usage</h2>
 
         <h3>Example: Grid</h3>
         <div class="example-grid">
-        <syntax-highlight>
-          <section class="grid grid-cols-3">
-            <div>1</div>
-            <div>2</div>
-            <div>3</div>
-            <div>4</div>
-            <div>5</div>
+          <syntax-highlight>
+            <section class="grid grid-cols-3">
+              <div>1</div>
+              <div>2</div>
+              <div>3</div>
+              <div>4</div>
+              <div>5</div>
+            </section>
+          </syntax-highlight>
+          <section class="example">
+            <section class="grid grid-cols-3 gap-2">
+              <div class="p-16 bg-gray-200 rounded-4">1</div>
+              <div class="p-16 bg-gray-200 rounded-4">2</div>
+              <div class="p-16 bg-gray-200 rounded-4">3</div>
+              <div class="p-16 bg-gray-200 rounded-4">4</div>
+              <div class="p-16 bg-gray-200 rounded-4">5</div>
+            </section>
           </section>
-        </syntax-highlight>
-        <section class="example">
-          <section class="grid grid-cols-3 gap-2">
-            <div class="p-16 bg-gray-200 rounded-4">1</div>
-            <div class="p-16 bg-gray-200 rounded-4">2</div>
-            <div class="p-16 bg-gray-200 rounded-4 ">3</div>
-            <div class="p-16 bg-gray-200 rounded-4 ">4</div>
-            <div class="p-16 bg-gray-200 rounded-4 ">5</div>
-          </section>
-        </section>
         </div>
         <p>3 column grid.</p>
 
         <h3>Example: Responsive 2/3, 1/3 Grid</h3>
         <syntax-highlight>
           <section class="md:grid md:grid-cols-3">
-              <div class="md:col-span-2">1</div>
-              <div>2</div>
+            <div class="md:col-span-2">1</div>
+            <div>2</div>
           </section>
         </syntax-highlight>
         <section class="example">
           <section class="md:grid md:grid-cols-3 gap-2">
-              <div class="p-16 bg-gray-200 rounded-4 md:col-span-2 mb-2 md:mb-0">1</div>
-              <div class="p-16 bg-gray-200 rounded-4 mb-2 md:mb-0">2</div>
+            <div class="p-16 bg-gray-200 rounded-4 md:col-span-2 mb-2 md:mb-0">1</div>
+            <div class="p-16 bg-gray-200 rounded-4 mb-2 md:mb-0">2</div>
           </section>
         </section>
         <p>
@@ -73,7 +73,7 @@
           <code>display:block</code>
           behavior(Single column) and DOM order on smaller devices, dividing it into 2/3 and 1/3 from medium
           devices and up. The
-          <a href="https://tailwindcss.com/docs/responsive-design">responsive prefix</a>
+          <a href="https://tailwindcss.com/pages/responsive-design">responsive prefix</a>
           <code>md:</code>
           turns the classes on above the medium breakpoint.
         </p>
@@ -110,7 +110,7 @@
         <p>
           Simple grid, could be a result page in card view, could be recommendations. Children are divided
           into a same size grid, the
-          <a href="https://tailwindcss.com/docs/responsive-design">responsive prefixes</a>
+          <a href="https://tailwindcss.com/pages/responsive-design">responsive prefixes</a>
           (md:, lg:) give them 2 columns on medium, and 3 on large devices.
         </p>
       </main>
