+++ pages/spacing.html
@@ -9,9 +9,9 @@
 
         <h2>References</h2>
         <ul>
-          <li><a href="https://tailwindcss.com/docs/margin">Margin in Tailwind</a></li>
-          <li><a href="https://tailwindcss.com/docs/padding">Padding in Tailwind</a></li>
-          <li><a href="https://tailwindcss.com/docs/responsive-design">Responsive prefix</a></li>
+          <li><a href="https://tailwindcss.com/pages/margin">Margin in Tailwind</a></li>
+          <li><a href="https://tailwindcss.com/pages/padding">Padding in Tailwind</a></li>
+          <li><a href="https://tailwindcss.com/pages/responsive-design">Responsive prefix</a></li>
         </ul>
 
         <h2>Fabric config</h2>
@@ -19,7 +19,7 @@
 
         <p>
           Available sizes from
-          <a href="https://github.com/fabric-ds/css/blob/main/packages/tailwind-config/tailwind.config.js">
+          <a href="https://github.com/fabric-ds/css/blob/main/src/tailwind/tailwind.config.js">
             tailwind-config
           </a>
           :
@@ -33,7 +33,7 @@
         <h2>Common usage</h2>
         <p>
           These can be used in a number of ways as described in more detail in the
-          <a href="https://tailwindcss.com/docs/padding">Tailwind documentation</a>
+          <a href="https://tailwindcss.com/pages/padding">Tailwind documentation</a>
           . The syntax is exactly the same for margins and paddings , just m instead of p.
         </p>
