+++ src/components/slider.css
