+++ src/scripts/replaceTailwindVariables.mjs
