+++ src/tailwind/README.md
@@ -3,7 +3,7 @@
 Fabric's [Tailwind](https://tailwindcss.com/) configuration.
 
 You can use this as a base if you want to customize the generated Tailwind classes. See the
-[Tailwind presets documentation](https://tailwindcss.com/docs/presets) for more information.
+[Tailwind presets documentation](https://tailwindcss.com/pages/presets) for more information.
 
 **Note that for FINN pages, utilities should never be removed from this preset. Components and podlets may
 assume that the utilities defined here are available. You can however, include further utilities if you
