+++ tailwind.config.js
@@ -1,4 +1,4 @@
-const preset = require('./packages/tailwind-config');
+const preset = require('./src/tailwind/tailwind.config');
 
 module.exports = {
     presets: [preset],
